CPU/GPU speed tests finished

+++ acmodel/matrix.py
@@ -7,8 +7,8 @@ import numpy as np
 import torch
 #device = torch.device("cuda:0")
 device = torch.device("cpu")
-#dtype = torch.float
-dtype = torch.double
+dtype = torch.float
+#dtype = torch.double
 
 #x = torch.rand(2, 3, device=device, dtype=dtype)
 #print(x)
